+++ tests/lib.rs
@@ -12,9 +12,9 @@ fn test_hotp() {
 #[test]
 fn test_totp() {
     let auth = otpauth::TOTP::new("python");
-    let timestamp1 = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs() as usize;
+    let timestamp1 = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
     let code = auth.generate(30, timestamp1);
-    let timestamp2 = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs() as usize;
+    let timestamp2 = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
     assert!(auth.verify(code, 30, timestamp2));
     assert!(!auth.verify(123456, 30, timestamp2));
     assert!(!auth.verify(1234567, 30, timestamp2));
